[tests] Changed some member types in types sample

+++ tests/samples/types/modB.f90
@@ -3,19 +3,19 @@ MODULE modB
   IMPLICIT NONE
 
   TYPE :: type0
-    INTEGER :: member
+    REAL :: member
   END TYPE type0
 
   TYPE :: type1
-    INTEGER :: member
+    REAL :: member
   END TYPE type1
 
   TYPE :: type2
-    INTEGER :: member
+    REAL :: member
   END TYPE type2
 
   TYPE :: type3
-    INTEGER :: member
+    REAL :: member
   END TYPE type3
 
 END MODULE modB
